requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data processing
numpy>=1.24.0
pillow>=9.0.0
opencv-python>=4.7.0
albumentations>=1.3.0

# Training utilities
tqdm>=4.65.0
tensorboard>=2.13.0
wandb>=0.15.0
pytorch-lightning>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
einops>=0.6.0

# Optional: Quality metrics
lpips>=0.1.4
pytorch-fid>=0.3.0

# Optional: Memory optimization
ninja>=1.11.0
torch-tb-profiler>=0.4.0 